requirements.txt
# Core dependencies
torch==2.1.2
transformers==4.36.2
datasets==2.16.1
accelerate==0.25.0
peft==0.7.1

# Data processing
pandas==2.1.4
numpy==1.24.3
python-dotenv==1.0.0
pyyaml==6.0.1

# Utilities
tqdm==4.66.2
click==8.1.7
pydantic==2.5.3

# Development tools
jupyter==1.0.0
ipython==8.19.0
pytest==7.4.3
black==23.12.1
flake8==6.1.0
